Done lots of logic work

+++ src/main/java/com/konzerra/selim_server/domain/review/dto/ReviewSaveDto.java
@@ -13,5 +13,5 @@ public class ReviewSaveDto {
     private String firstName;
     private String lastName;
     private String reviewText;
-    private Integer gateCategoryId;
+    private Long gateCategoryId;
 }
